web/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-gravatar2==1.4.4
django-hCaptcha==0.1.0
django-heroku==0.3.1
django-utils-six==2.0
django-widget-tweaks==1.4.9
fonttools==4.28.5
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiwisolver==1.3.2
msgpack==1.0.3
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing-extensions==4.1.1
urllib3==1.26.7
whitenoise==5.2.0
zope.interface==5.4.0
